pqdenoise: Simplify
There are a lot of functions here that can be immensely simplified with more ergonomic helper functions from the vs-iew suite. There's also a number of functions here that could be replaced in its entirely.

For now, I tried simplifying the first function in the module as a POC. If there's interest in me trying to simplify other functions, let me know in this PR and I may see if I have some time to do so.